requirements.txt
selenium>=4.0.0
python-docx>=0.8.11
webdriver-manager>=3.8.4
uvicorn
fastapi
pydantic